+++ web/app/pages/desktop/module.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import {Route} from 'react-router';
 
-import Desktop from './desktop.component.jsx';
-import PreventDoubleClick from './prevent-double-click.component.jsx';
-import WithResolves from './with-resolves.component.jsx';
-import WithParam from './with-param.component.jsx';
+import DesktopPage from './desktop.page.jsx';
+import PreventDoubleClickPage from './prevent-double-click.page.jsx';
+import WithResolvesPage from './with-resolves.page.jsx';
+import WithParamPage from './with-param.page.jsx';
 
 import {
   desktop as desktopHooks,
@@ -16,27 +16,27 @@ const routes = [
     key="1"
     name="desktop"
     path="/desktop"
-    component={Desktop}
+    component={DesktopPage}
     onEnter={desktopHooks.onEnter}
   />,
   <Route
     key="2"
     name="prevent-double-click"
     path="/prevent-double-click"
-    component={PreventDoubleClick}
+    component={PreventDoubleClickPage}
   />,
   <Route
     key="3"
     name="with-resolves"
     path="/with-resolves"
-    component={WithResolves}
+    component={WithResolvesPage}
     onEnter={withResolvesHooks.onEnter}
   />,
   <Route
     key="4"
     name="with-param"
     path="/with-param(/:p1)"
-    component={WithParam}
+    component={WithParamPage}
   />
 ];