+++ web/app/react/components.js
@@ -1,9 +1,0 @@
-import Application from './components/application.component.jsx';
-import Header from './components/header.component.jsx';
-import NotFound from './components/not-found.component.jsx';
-
-export {
-  Application,
-  Header,
-  NotFound
-};