+++ web/app/routes.jsx
@@ -8,11 +8,9 @@ import {
 import {Provider} from 'react-redux';
 import {syncHistoryWithStore} from 'react-router-redux';
 import store from './store/store';
-import {
-  Application,
-  NotFound
-} from './react/components';
-import Desktop from './pages/desktop/desktop.component.jsx';
+import Application from './react/components/application.component.jsx';
+import NotFoundPage from './react/components/not-found.page.jsx';
+import DesktopPage from './pages/desktop/desktop.page.jsx';
 import {routes as desktopRoutes} from './pages/desktop/module.jsx';
 
 let history = syncHistoryWithStore(browserHistory, store);
@@ -21,9 +19,9 @@ export default (
   <Provider store={store}>
     <Router history={history}>
       <Route path="/" component={Application}>
-        <IndexRoute component={Desktop} />
+        <IndexRoute component={DesktopPage} />
         {desktopRoutes}
-        <Route path="*" component={NotFound} />
+        <Route path="*" component={NotFoundPage} />
       </Route>
     </Router>
   </Provider>